feat(core): added input and moved component to proper place
