+++ libs/core/src/lib/menu/menu.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { MenuComponent } from './menu.component';
 import { MenuTitleDirective } from './directives/menu-title.directive';
-import { MenuInteractiveComponent } from './directives/menu-interactive.component';
+import { MenuInteractiveComponent } from './menu-interactive.component';
 import { MenuSeparatorDirective } from './menu-separator.directive';
 import { MenuAddonDirective } from './directives/menu-addon.directive';
 import { MenuItemComponent, SubmenuComponent } from './menu-item/menu-item.component';
@@ -17,6 +17,7 @@ import { PortalModule } from '@angular/cdk/portal';
 import { ToggleButtonDirective } from './directives/toggle-button.directive';
 import { SegmentedButtonHeaderDirective } from './directives/segmented-button/segmented-button-header.directive';
 import { SegmentedButtonOptionDirective } from './directives/segmented-button/segmented-button-option.directive';
+import { MenuItemInputDirective } from './directives/menu-item-input.directive';
 
 @NgModule({
     imports: [
@@ -37,7 +38,8 @@ import { SegmentedButtonOptionDirective } from './directives/segmented-button/se
         MenuTriggerDirective,
         ToggleButtonDirective,
         SegmentedButtonHeaderDirective,
-        SegmentedButtonOptionDirective
+        SegmentedButtonOptionDirective,
+        MenuItemInputDirective
     ],
     exports: [
         MenuComponent,
@@ -53,7 +55,8 @@ import { SegmentedButtonOptionDirective } from './directives/segmented-button/se
         GlyphMenuAddonDirective,
         ToggleButtonDirective,
         SegmentedButtonHeaderDirective,
-        SegmentedButtonOptionDirective
+        SegmentedButtonOptionDirective,
+        MenuItemInputDirective
     ],
     providers: [DynamicComponentService]
 })
